refactor: move subject assignments to separated project
