+++ Source/Modules/Iwentys.Modules.Assignments/Queries/CreateAssignment.cs
@@ -5,10 +5,12 @@ using System.Threading.Tasks;
 using Iwentys.Domain.Assignments;
 using Iwentys.Domain.Assignments.Models;
 using Iwentys.Domain.Study;
+using Iwentys.Infrastructure.Application;
 using Iwentys.Infrastructure.DataAccess;
+using Iwentys.Modules.Assignments.Dtos;
 using MediatR;
 
-namespace Iwentys.Infrastructure.Application.Controllers.Assignments
+namespace Iwentys.Modules.Assignments.Queries
 {
     public static class CreateAssignment
     {
