+++ Source/Modules/Iwentys.Modules.Gamification/Leaderboard/Dtos/StudyLeaderboardRowDto.cs
