+++ Source/Modules/Iwentys.Modules.Gamification/Leaderboard/Queries/CourseRatingForceRefresh.cs
@@ -48,11 +48,34 @@ namespace Iwentys.Infrastructure.Application.Controllers.Leaderboard
                     .GetStudentActivities(new StudySearchParametersDto { CourseId = request.CourseId })
                     .ToList();
                 
-                List<CourseLeaderboardRow> newRows = CourseLeaderboardRow.Create(request.CourseId, result, oldRows);
+                List<CourseLeaderboardRow> newRows = Create(request.CourseId, result, oldRows);
 
                 _context.CourseLeaderboardRows.RemoveRange(oldRows);
                 _context.CourseLeaderboardRows.AddRange(newRows);
                 return new Response();
+            }
+
+            public static List<CourseLeaderboardRow> Create(int courseId, List<SubjectActivity> rows, List<CourseLeaderboardRow> oldRows)
+            {
+                Dictionary<int, int> mapToOld = oldRows.ToDictionary(v => v.StudentId, v => v.Position);
+
+                return rows
+                    .GroupBy(r => r.StudentId)
+                    .Select(g => new StudyLeaderboardRowDto(g.ToList()))
+                    .OrderByDescending(a => a.Activity)
+                    .Take(50)
+                    .OrderByDescending(r => r.Activity)
+                    .Select((r, position) => CreateRow(r, courseId, position, mapToOld))
+                    .ToList();
+            }
+
+            private static CourseLeaderboardRow CreateRow(StudyLeaderboardRowDto row, int courseId, int position, Dictionary<int, int> mapToOld)
+            {
+                int? oldPosition = null;
+                if (mapToOld.TryGetValue(row.Student.Id, out var value))
+                    oldPosition = value;
+
+                return new CourseLeaderboardRow { Position = position + 1, CourseId = courseId, StudentId = row.Student.Id, OldPosition = oldPosition };
             }
         }
     }
