+++ Source/Modules/Iwentys.Modules.Study/Study/Dtos/GroupProfileResponseDto.cs
@@ -5,7 +5,7 @@ using System.Linq.Expressions;
 using Iwentys.Domain.Study;
 using Iwentys.Domain.Study.Models;
 
-namespace Iwentys.Infrastructure.Application.Controllers.Study.Dtos
+namespace Iwentys.Modules.Study.Study.Dtos
 {
     public record GroupProfileResponseDto
     {
