+++ Source/Modules/Iwentys.Modules.Study/Study/Dtos/StudentActivityInfoDto.cs
@@ -2,7 +2,7 @@
 using Iwentys.Common.Tools;
 using Iwentys.Domain.Study;
 
-namespace Iwentys.Infrastructure.Application.Controllers.Study.Dtos
+namespace Iwentys.Modules.Study.Study.Dtos
 {
     public record StudentActivityInfoDto
     {
