+++ Source/Modules/Iwentys.Modules.Study/Study/Dtos/SubjectActivityInfoResponseDto.cs
@@ -1,6 +1,6 @@
 ﻿using Iwentys.Domain.Study;
 
-namespace Iwentys.Infrastructure.Application.Controllers.Study.Dtos
+namespace Iwentys.Modules.Study.Study.Dtos
 {
     public record SubjectActivityInfoResponseDto
     {
