+++ Source/Modules/Iwentys.Modules.Study/Study/Dtos/SubjectProfileDto.cs
@@ -1,6 +1,6 @@
 ﻿using Iwentys.Domain.Study;
 
-namespace Iwentys.Infrastructure.Application.Controllers.Study.Dtos
+namespace Iwentys.Modules.Study.Study.Dtos
 {
     public record SubjectProfileDto
     {
