+++ Source/Modules/Iwentys.Modules.Study/Study/Queries/GetSubjectsByGroupId.cs
@@ -3,13 +3,13 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Iwentys.Domain.Study.Models;
-using Iwentys.Infrastructure.Application.Controllers.Study.Dtos;
 using Iwentys.Infrastructure.Application.Repositories;
 using Iwentys.Infrastructure.DataAccess;
+using Iwentys.Modules.Study.Study.Dtos;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace Iwentys.Infrastructure.Application.Controllers.Study
+namespace Iwentys.Modules.Study.Study.Queries
 {
     public class GetSubjectsByGroupId
     {
