+++ Source/Modules/Iwentys.Modules.Study/Study/StudyGroupController.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Iwentys.Infrastructure.Application;
-using Iwentys.Infrastructure.Application.Controllers.Study;
-using Iwentys.Infrastructure.Application.Controllers.Study.Dtos;
+using Iwentys.Modules.Study.Study.Dtos;
+using Iwentys.Modules.Study.Study.Queries;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
