+++ Source/Modules/Iwentys.Modules.Study/StudyMappingProfile.cs
@@ -1,0 +1,11 @@
+﻿using AutoMapper;
+
+namespace Iwentys.Modules.Study
+{
+    public class StudyMappingProfile : Profile
+    {
+        public StudyMappingProfile()
+        {
+        }
+    }
+}
