+++ Source/Modules/Iwentys.Modules.SubjectAssignments/Dtos/SubjectAssignmentDto.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Iwentys.Domain.AccountManagement.Dto;
 using Iwentys.Domain.SubjectAssignments.Enums;
 
-namespace Iwentys.Infrastructure.Application.Modules.SubjectAssignments.Dtos
+namespace Iwentys.Modules.SubjectAssignments.Dtos
 {
     public class SubjectAssignmentDto
     {
