+++ Source/Modules/Iwentys.Modules.SubjectAssignments/Dtos/SubjectAssignmentSubmitDto.cs
@@ -3,7 +3,7 @@ using Iwentys.Domain.Study.Models;
 using Iwentys.Domain.SubjectAssignments;
 using Iwentys.Domain.SubjectAssignments.Enums;
 
-namespace Iwentys.Infrastructure.Application.Modules.SubjectAssignments.Dtos
+namespace Iwentys.Modules.SubjectAssignments.Dtos
 {
     public class SubjectAssignmentSubmitDto
     {
