+++ Source/Modules/Iwentys.Modules.SubjectAssignments/Iwentys.Modules.SubjectAssignments.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Iwentys.Modules.Study\Iwentys.Modules.Study.csproj" />
+  </ItemGroup>
+
 </Project>
