+++ Source/Modules/Iwentys.Modules.SubjectAssignments/MentorScope/MentorSubjectAssignmentSubmitController.cs
@@ -1,15 +1,14 @@
-﻿using System;
-using Iwentys.Infrastructure.Application.Modules.SubjectAssignments.MentorScope.Queries;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using Iwentys.Domain.SubjectAssignments.Models;
+using Iwentys.Infrastructure.Application;
+using Iwentys.Modules.SubjectAssignments.Dtos;
+using Iwentys.Modules.SubjectAssignments.MentorScope.Queries;
+using Iwentys.Modules.SubjectAssignments.StudentScope.Queries;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
-using System.Collections.Generic;
-using Iwentys.Common.Tools;
-using Iwentys.Infrastructure.Application.Modules.SubjectAssignments.StudentScope.Queries;
-using Iwentys.Infrastructure.Application.Modules.SubjectAssignments.Dtos;
-using Iwentys.Domain.SubjectAssignments.Models;
 
-namespace Iwentys.Infrastructure.Application.Modules.SubjectAssignments.MentorScope
+namespace Iwentys.Modules.SubjectAssignments.MentorScope
 {
     [Route("api/subject-assignment/mentor/submit")]
     [ApiController]
