+++ Source/Modules/Iwentys.Modules.SubjectAssignments/MentorScope/Queries/DeleteSubjectAssignment.cs
@@ -3,11 +3,12 @@ using System.Threading.Tasks;
 using AutoMapper;
 using Iwentys.Domain.AccountManagement;
 using Iwentys.Domain.SubjectAssignments;
-using Iwentys.Infrastructure.Application.Modules.SubjectAssignments.Dtos;
+using Iwentys.Infrastructure.Application;
 using Iwentys.Infrastructure.DataAccess;
+using Iwentys.Modules.SubjectAssignments.Dtos;
 using MediatR;
 
-namespace Iwentys.Infrastructure.Application.Modules.SubjectAssignments.MentorScope.Queries
+namespace Iwentys.Modules.SubjectAssignments.MentorScope.Queries
 {
     public static class DeleteSubjectAssignment
     {
