+++ Source/Modules/Iwentys.Modules.SubjectAssignments/MentorScope/Queries/UpdateSubjectAssignment.cs
@@ -4,11 +4,12 @@ using AutoMapper;
 using Iwentys.Domain.AccountManagement;
 using Iwentys.Domain.SubjectAssignments;
 using Iwentys.Domain.SubjectAssignments.Models;
-using Iwentys.Infrastructure.Application.Modules.SubjectAssignments.Dtos;
+using Iwentys.Infrastructure.Application;
 using Iwentys.Infrastructure.DataAccess;
+using Iwentys.Modules.SubjectAssignments.Dtos;
 using MediatR;
 
-namespace Iwentys.Infrastructure.Application.Modules.SubjectAssignments.MentorScope.Queries
+namespace Iwentys.Modules.SubjectAssignments.MentorScope.Queries
 {
     public static class UpdateSubjectAssignment
     {
