+++ Source/Tests/Iwentys.Tests/TestCaseContexts/StudyTestCaseContext.cs
@@ -7,9 +7,9 @@ using Iwentys.Domain.Study.Models;
 using Iwentys.Domain.SubjectAssignments;
 using Iwentys.Domain.SubjectAssignments.Enums;
 using Iwentys.Domain.SubjectAssignments.Models;
-using Iwentys.Infrastructure.Application.Controllers.Study.Dtos;
 using Iwentys.Infrastructure.DataAccess.Seeding.FakerEntities;
 using Iwentys.Infrastructure.DataAccess.Seeding.FakerEntities.Study;
+using Iwentys.Modules.Study.Study.Dtos;
 
 namespace Iwentys.Tests.TestCaseContexts
 {
